Address #185 by using "-s" option with ssh-agent to ensure Bourne-compatible output.
This allows simplification of write_pidfile(), as we don't need to adapt csh-style pidfile output.
It also means that "eval $pidfile_out" is now equivalent to "$(eval catpidf_shell sh)" in situations
where the pidfile was written. A defensive check against "setenv" in the pidfile was added as a
precaution against situations where the "-s" option may not work.

In addition, a compatibility section was added to the man page. Ssh.com support was officially
dropped, as it is not going to support the "ssh-agent -s" option, and this variant of ssh is not
likely to be encountered on systems that are still in use in 2025.

Other improvements:
* Add .gitattributes and .vscode/settings.json to prepare for Windows 11 bash git testing.
* Minor ShellCheck fixes for release scripts.
* Comment typo fix.